product-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mongoengine==0.27.0
python-decouple==3.8
django-cors-headers==4.3.1
drf-spectacular==0.26.5
gunicorn==21.2.0
pymongo==4.6.1
dnspython==2.4.2
prometheus-client==0.19.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
mongomock==4.1.2
